emac: Enable sleep support for EMAC RXC clock
EMAC RXC clock is always sourced from PHY. If the GPIO
used for RXC clock is multiplexed, there will be interrupt
line routed towards GIC and it will be always enabled.
As a result, device will never go to low power state.
Make a change to disable GPIO pin used for RXC clock
when Defenderice goes to suspend and enable it back on resume.

Change-Id: I83cab3f77ee05b12004a28f68529566de5fd1faa
Signed-off-by: Chaitanya Pratapa <cpratapa@codeaurora.org>